% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_corporate.R
\name{theme_corporate}
\alias{theme_corporate}
\title{Ggplot2 theme}
\usage{
theme_corporate()
}
\value{
A theme for ggplot2.
}
\description{
Ggplot2 theme
}
\examples{
ggplot(mtcars, aes(mpg, wt, colour = factor(cyl))) + geom_point(size=4) +theme_corporate()
}
